DOCS: Added undercornertill* and updated aroundcornertill*
Now undercornertill* and aroundcornertill* will take 4 arguments. First
is the vector for magnitude, second is the pivot point, third is the
till point, fourth is the vector from till point. The edge will be
extended or clipped till the intersection of the this vector and the
previously marked FCBTMP -- SCBTMP.